requirements.txt
langchain
langchain-openai
openai
python-dotenv
yfinance  # for stock data
pandas
numpy